All works

+++ backendSpring/src/main/java/org/java/app/photoalbum/controller/PhotoController.java
@@ -7,6 +7,9 @@ import org.java.app.photoalbum.db.pojo.Photo;
 import org.java.app.photoalbum.db.serv.CategoryService;
 import org.java.app.photoalbum.db.serv.PhotoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -16,6 +19,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jakarta.validation.Valid;
@@ -29,28 +33,36 @@ public class PhotoController {
 	@Autowired
 	private CategoryService categoryService;
 	
-	private String savePhoto(Photo photo, BindingResult bindingResult, Model model, 
-			RedirectAttributes ra, boolean isNew) {
-		if (bindingResult.hasErrors()) {
-			model.addAttribute("categories", categoryService.findAll());
-			return "photos/photo-create";
-		}
-		else {
-			ra.addFlashAttribute("updateMessage", "Foto " + (isNew ? 
-															"aggiunta " :
-															"modificata ")
-														+ "correttamente!");
-			photoService.save(photo);
-			return "redirect:/photos/" + photo.getId();
-
-		}}
+	private boolean isSuperAdmin(UserDetails user) {
+		return user.getAuthorities().stream().anyMatch(auth -> "SUPER_ADMIN".equals(auth.getAuthority()));
+		
+	}
+	
+	private boolean sameUser(UserDetails userPhoto, UserDetails loggedUser) {
+		return userPhoto.getUsername().equals(loggedUser.getUsername()) 
+				&& userPhoto.getPassword().equals(loggedUser.getPassword());
+	}
 	
 	@GetMapping public String getIndex(Model model,
-			@RequestParam(required = false) String title) {
+			@RequestParam(required = false) String title, 
+			@AuthenticationPrincipal org.java.app.photoalbum.auth.pojo.User user) {
 		
-		List<Photo> photos = title == null ?
-							 photoService.findAll():
-							 photoService.findByTitle(title);
+		List<Photo> photos = null;
+		
+		if (!isSuperAdmin(user)) {
+		 photos = title == null ?
+							 photoService.findAllByUser(user):
+							 photoService.findAllByUserWithTitle(user, title);
+		
+		}
+		
+		else {
+			
+			 photos = title == null ?
+					 photoService.findAll():
+					 photoService.findByTitle(title);
+			
+		}
 		
 		model.addAttribute("photos", photos);
 		model.addAttribute("title", title);
@@ -60,7 +72,9 @@ public class PhotoController {
 	}
 	
 	@GetMapping("/{id}")
-	public String getShow(@PathVariable int id, Model model) {
+	public String getShow(@PathVariable int id, Model model, @AuthenticationPrincipal UserDetails user) {
+		
+		
 		
 		Optional<Photo> otpPhoto = photoService.findById(id);
 		if (otpPhoto.isEmpty()) {
@@ -68,6 +82,13 @@ public class PhotoController {
 		}
 		Photo photo = otpPhoto.get();
 		model.addAttribute("photo", photo);
+		
+		UserDetails userPhoto = photo.getUser();
+
+		if (!sameUser(userPhoto, user) && (!isSuperAdmin(user)))
+		    throw new ResponseStatusException(HttpStatus.FORBIDDEN); 
+
+		
 		
 		return "photos/photo-show";
 	}
@@ -83,17 +104,35 @@ public class PhotoController {
 	
 	@PostMapping("/create")
 	public String storePhoto(@Valid @ModelAttribute Photo photo, BindingResult bindingResult, Model model,
-			RedirectAttributes ra) {
-		return savePhoto(photo, bindingResult, model, ra, true);
-	}
+			RedirectAttributes ra, @AuthenticationPrincipal UserDetails user) {
+		
+	
+		if (bindingResult.hasErrors()) {
+			model.addAttribute("categories", categoryService.findAll());
+			return "photos/photo-create";
+		}
+		else {
+			ra.addFlashAttribute("updateMessage", "Foto aggiunta correttamente!");
+			photo.setUser((org.java.app.photoalbum.auth.pojo.User) user);
+			photoService.save(photo);
+			return "redirect:/photos/" + photo.getId();
+			
+	}}
 	
 	@GetMapping("/update/{id}")
-	public String photoUpdate(@PathVariable int id, Model model) {
+	public String photoUpdate(@PathVariable int id, Model model, @AuthenticationPrincipal UserDetails user) {
+		Photo fullInfoPhoto = photoService.findById(id).get();
+		org.java.app.photoalbum.auth.pojo.User userPhoto = fullInfoPhoto.getUser();
+		if (!sameUser(userPhoto, user) && (!isSuperAdmin(user))) 
+		    throw new ResponseStatusException(HttpStatus.FORBIDDEN); 
+		
+		
 		Optional<Photo> optPhoto = photoService.findById(id);
 		if (optPhoto.isEmpty()) {
 			return "redirect:/photos";
 		}
 		Photo photo = optPhoto.get();
+
 		model.addAttribute("photo", photo);
 		model.addAttribute("categories", categoryService.findAll());
 
@@ -103,17 +142,41 @@ public class PhotoController {
 	
 	@PostMapping("/update/{id}")
 	public String updatePhoto(@Valid @ModelAttribute Photo photo, BindingResult bindingResult, Model model,
-			RedirectAttributes ra) {
-		return savePhoto(photo, bindingResult, model, ra, false);		
-	}
+			RedirectAttributes ra, @AuthenticationPrincipal UserDetails user) {
+		
+		Photo fullInfoPhoto = photoService.findById(photo.getId()).get();
+		org.java.app.photoalbum.auth.pojo.User userPhoto = fullInfoPhoto.getUser();
+		if (!sameUser(userPhoto, user) && (!isSuperAdmin(user))) 
+		    throw new ResponseStatusException(HttpStatus.FORBIDDEN); 
+		
+		if (bindingResult.hasErrors()) {
+			model.addAttribute("categories", categoryService.findAll());
+			return "photos/photo-create";
+		}
+		else {
+			
+			ra.addFlashAttribute("updateMessage", "Foto aggiunta correttamente!"); 
+			photo.setUser(userPhoto);
+			photoService.save(photo);
+			return "redirect:/photos/" + photo.getId();
+		}}
 	
 	@PostMapping("/delete/{id}")
-	public String deletePizza(@PathVariable int id, RedirectAttributes ra) {
+	public String deletePizza(@PathVariable int id, RedirectAttributes ra, @AuthenticationPrincipal UserDetails user) {
 		Optional<Photo> optPhoto = photoService.findById(id);
 		if (optPhoto.isEmpty()) {
 			return "redirect:/photos";
 		}
+		
 		Photo photo = optPhoto.get();
+
+		org.java.app.photoalbum.auth.pojo.User userPhoto = photo.getUser();
+
+		if (!sameUser(userPhoto, user) && (!isSuperAdmin(user)))
+		    throw new ResponseStatusException(HttpStatus.FORBIDDEN); 
+
+		
+		
 		ra.addFlashAttribute("updateMessage", "Foto con ID: " + photo.getId() + " (" + photo.getTitle() + ") cancellata");
 		photoService.delete(photo);
 		return "redirect:/photos";
